requirements.txt
# ── Core Framework ────────────────────────────────────────────────────
flask>=3.0
flask-sqlalchemy>=3.1
flask-login>=0.6
flask-bcrypt>=1.0
flask-migrate>=4.0
flask-limiter>=3.5
python-dotenv>=1.0
Werkzeug>=3.0

# ── RAG Pipeline ──────────────────────────────────────────────────────
langchain>=0.3
langchain-community>=0.3
langchain-google-genai>=2.0
langchain-text-splitters>=0.3
faiss-cpu>=1.8
pypdf>=4.0

# ── Gemini API ────────────────────────────────────────────────────────
google-generativeai>=0.8

# ── Production Server ────────────────────────────────────────────────
gunicorn>=22.0

# ── Database (uncomment for PostgreSQL) ──────────────────────────────
# psycopg2-binary>=2.9
